Reverting start file for MBLAZE and PPC440
Reverted start files (crt0.S) as I found out I had the
wrong preprocessor (i.e. wasn't included in the build) and
when I did include it properly, was having trouble linking
the symbols (e.g. load_my_table, init_PR_data, etc.) as that
is not defined for non-heterogeneous applications. Since the
last stages of the hcompile, calls jam, the linker complains
as it can't find these symbols for non-heterogeneous applications.

I can't think of a 'clean' way of fixing this and for now, will
manually called these init subroutines at the beginning of main
when generating data. Not a final solution, but this optimization
is also not a priority and probably doesn't affect the results
(especially for a long running program).