HCOMPILE CAN PREFETCH ACCELERATORS AND DECLARE IN SLAVE TABLE
Hcompile can now read from the system hw definition file, and
determine (based on a list of supported accelerators) which
processors are connected to which accelerators. I assume an
accelerator is connected to only one processor.

Also updating the IDs for accelerators. Need to go back at
some point and generalize it a bit. In some ways, the polymorphic
functions are hardcoded to use a specific version of sort (or a
vector function) when hcompile pulls specific types of accelerators
(i.e. bubblesort, and vectoradd and vectormul).