AXI TIMER INTEGRATED WITHIN HTHREADS
AXI Timer 2.0 has been integrated in the hthreads boot process.
Hthreads configure this timer, assumed to be seen by all, but
configured by one (host), as a 64-bit free-running timer. Users
can continue to use hthread_time_get(). A slight change is that
the hthread_time_*sec calls now return a double instead of a float
to account for the 64-bit width.