Build scripts can generate both dcp and IP catalog formats for accelerators.
Ran into an issue where exporting the design twice (once to create dcp files
for PR flow and the other for Vivado's IP catalog). When doing this, it
seems that the exported IP catalog would not work. I had to change
the run_hls.tcl files for each accelerator to create two solutions
within a project: one for PR flow (if defined), and the other for
the IP catalog when building systems.

Also removed dcp files from version control.

Renamed top-level design for matrix multiply

Fixed a typo in sdk_config.sh