Modified TCB structure: Added execution time field
Modified the TCB structure to include execution time. Modified
the slave's hardware kernel to record this automatically when
exiting the thread. Also, modified _bootstrap_thread to achieve
the same results when scheduling software threads on the host.
I have done the same for _boostrap_thread for slave processors,
but slave processors currently do no support the creation of software
threads.

The catch, however, is it assumes you read this field once joining
on a thread and before creating a new thread. The call, hthread_join
returns the return_value field from the TCB on that call and nothing else.
Once creating a new thread after joining on threads, these threads
and their respective structures are recycled. In order not to deviate
from the Pthread spec, I created my own join call that checks whether
a thread has exited, read the execution time, and then proceed with
joining on that thread.