Incorrect flag option removed"
Removed incorrect flag that was causing incorrect compiler
output. The flag

-fno-math-errno

This shouldn't be used with any -O option if I am
intepreting the gcc manual correctly:

"This option should never be turned on by any -O option since
it can result in incorrect output for programs which depend on
an exact implementation of IEEE or ISO rules/specifications
for math functions.

The default is -fmath-errno."