Hcompile: Replaced echo commands with printf
Replaced echo commands with the portable printf command. Hcompile
was relying on a non-coformant POSIX version of echo (in the bash
shell). Hcompile required your "/bin/sh" to be pointing to
bash instead of dash (or something else that conforms to the POSIX
standard of of echo). This is fixed now.

The platform build system needs to be updated with this change.