Saving final days of work for dissertation.
As of July 23rd 2016, this is the latest copy of Hthreads (sprinkled with some of
my dissertation-specific modifications). So it is experimental. I do not know if
there is future planned development on Hthreads but if so, the person or persons
should start with this git repo.

The last person to work on Hthreads besides myself is Alborz. Towards the end
of our graduation date, we did not get the opportunity to merge our branches.
As far as I am aware of, his branch had a few accelerators adopted from the
CGRA work currently being investigated by the research lab (2016).

-Eugene